m4: fix c++ check for -f flags
Both gcc and clang generate warnings for unsupported -f$FLAG by
default, meaning that the previous implementation would consider
unsupported flags as supported (as a warning is not an error that
fails compilation).

The addition of -Werror treats warnings as errors, and will prevent
false-positives in terms of -f$FLAG support.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>