gui/qt: simple_preferences: combine ppsz_language and ppsz_language_map
Having things as two separate arrays is very hard to maintain as one
entry in ppsz_language must correctly correspond to one element in
ppsz_language_text.

These changes refactors the implementation so that a single array is
used, where each element has two members so that it is easier to see
what corresponds to what.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>