misc/text_style: text_segment_Copy: prevent memory-leak
If text_segment_New is successful at least once, and then unsuccessful
two times in a row we would lose the pointer to the first allocated
object (meaning that it would leak).

This patch fixes the issue by aborting the loop as soon as we
encounter an error (if any).

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>