input/control: INPUT_GET_SEEKPOINTS: prevent out-of-bounds reads
The previous implementation would potentially read the "element" right
after the last one in the subsequently used array (as priv->i_title
denotes the number of titles present).

These changes also include a check to make sure that var_GetInteger(
p_input, "title" ) does not result in a negative value (as such can be
used to read from outside the array, but on the other side).

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>