direct3d11: fix HDR rendering on LDR screens
The wide contrast provided by HDR results in washed out colors when displayed
on an LDR (non-HDR) screen. To fix this we need to do some tone mapping in
linear space. The tone mapping used here comes from [1].

The pixel shader source is constructed based on the video input characteristics.
When the source is LDR no tone mapping or transfer adjustment is done.

For now we assume the screen is using sRGB transfer.

1: http://filmicworlds.com/blog/filmic-tonemapping-operators/

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>