access/satip: fix null-dereference on missing host
The module assumes that the vlc_url_t populated by vlc_UrlParse always
have an entry in vlc_url_t::psz_host, which is not the case for
something such as "satip://.invalid:80".

These changes also make sure that we do not accept a URL where an
empty host is specified, as this is not accepted by the SAT>IP
specification.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>