audiotrack: test if encodings are supported
There is no way to know if a device supports a specific codec using the
ENCODING_IEC61937 audio format.

There is an API for that, starting Android 21. But unfortunately, you need a
Context and to implement an interface. This is not possible via JNI (except if
we depend on a proxy Java class).

Therefore, we let the Android LibVLC listening to audio devices events and
report new supported encodings via libvlc_audio_output_device_set().