demux/adaptive: DOMParser: retain root element if empty
An XML file containing a single root element without children would,
given the previous implementation, have its only element ignored.
Resulting in a memory-leak of the node in question, while also (more
importantly) making it impossible to access the associated data.

As XML does not allow documents such as the below (only a single root
is allowed):

   <?xml version="1.0">
   <ill-formed />
   <ill-formed></ill-formed>

Simply checking to see so that we are not popping away all our tags
are sufficient in order to fix this bug. The changes also make sure
that we do not invoke std::stack<...>::pop on an empty container
(which is undefined-behavior).

fixes: #18122

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>