misc/gnutls: do not declare variable if not needed
The variable in question is only needed within the preprocessor-if
that follows, as such it makes more sense to declare only declare it
if that is actually reached (while also supressing a warning
diagnostic related to the variable being unused).

An empty statement has been added after the label in order to comply
with the ISO C Standard, which mandates that a label preceedes a
statement (which a variable declaration is not).

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>