text_renderer/freetype: FontConfig_GetFamily: simplify bold/italic flags
The previous implementation would result in a diagnostic from gcc due
to it being unable to prove that b_bold and b_italic was always
initialized before usage.

These changes get rid of the diagnostic, while also simplifying the
implementation. A comment was added to make the purpose of the loop
clear.

Signed-off-by: Hugo Beauzée-Luyssen <hugo@beauzee.fr>