preparser: post-pone event until after art fetching is complete
Since one can request art to be fetched through
libvlc_media_parse_with_options, one would expect the event
originating from this request to be sent upon the completion of all
requested operations (not just the preparsing). The alternative would
be to monitor the libvlc_MediaMetaChanged, hoping for an artwork URL
change, but this can't account for error nor timeout.

--

The changes introduced were written after a discussion with Hugo
Beauzée-Luyssen where he expressed that he would, if possible, be able to
post-pone the preparsing events until the art fetching is complete.

Post-poning the event fixes issues that are currently reproducible where
medialibrary is used, and art is missing (because it listens to the preparse
event in order to generate the database contents).

It can be viewed as a follow-up to the below rejected patch (with the
same goal in mind):

 - https://patches.videolan.org/patch/15810/

Signed-off-by: Hugo Beauzée-Luyssen <hugo@beauzee.fr>