demux/ty: Demux: fix comment related to "closed captions/XDS"
As a result of dbae0e3, a comment was incorrectly split up so that it
implies that each split applies to one particular value (which is not
correct).

These changes fix that by putting both entities into a single comment,
hopefully signalling that either "0x01" or "0x02" means "closed
captions/XDS" (and not each for one).

--

Thanks for Francois Cartegnie for pointing out the mistake:

 - https://mailman.videolan.org/pipermail/vlc-devel/2017-March/112291.html

Signed-off-by: Hugo Beauzée-Luyssen <hugo@beauzee.fr>