demux/playlist: fix mime-type matching
The previous implementation would inaccurately truncate the mime-type
check in case of parameters in the content-type string, which in turn
would lead to false-positive matches.

As the usage of CheckContentType is really meant to check the
mime-type of the content-type string, besides fixing the inaccurate
comparision, the function is also renamed to CheckMimeType.

fixes: #18143

Signed-off-by: Rémi Denis-Courmont <remi@remlab.net>