vout: update interlacing handling
Remove deinterlace from 'video-filter' vlc string variable, now handling it
with a boolean.

This fixes a bug that happened when deinterlacing a video and applying a new
filter: the deinterlace filter was not reapplied when rebuilding the filter
chain.

We couldn't retrieve this var from the GUI as the presence of this filter can
change at any time, considering that a video can be partially interlaced many
times.

It's also cleaner to handle the presence of the deinterlace filter without
using a VLC variable as we only need to know if it's there in the video output
module.

Signed-off-by: Thomas Guillem <thomas@gllm.fr>