ttml: remove bogus casts
ptrdiff_t is the result type for a substraction of pointers, not the
operands type for the substraction. The cast made no sense.

To convert pointers to signed integers, intptr_t shall be used. But
there is no point in casting here: equality and inequality are always
well-defined for valid and/or NULL pointers.