ttml: remove nonsensical pointer cast
The difference between the integer representation of two pointers has
no particular meaning in general (the result is platform-dependent),
other than zero meaning equality. It is only really useful when dealing
with data alignment.

In this case, both operands point into the same object, so the pointer
difference is well-defined anyhow.