macOS: Assume in Xcode project that build exists
This is a vague assumption as the build folder could be called anything
actually or not be present at all, but for most users following the wiki
guide this will be true.
If it isn't, it will result in missing header errors, the same that
were present before this change.