event: remove recursive deletion
In theory, vlc_event_detach() can be called from within the event
handler. In practice, callers of vlc_event_detach() expect that the
event handler is not pending after the function returns. This would not
work if recursion actually occurred, it would lead to use-after-free.

This removes recursion, including memory allocation, copying and missing
error handling in event sending.