demux/playlist: xspf: fix skip of unknown <extension>
As a XML_READER_STARTELEM does not have a corresponding
XML_READER_ENDELEM if the tag is self-closing, the code responsible
for skipping unknown extension-tags would result in skips outside of
the target element.

These changes fixes the issue by making sure that we do not increase
the ignore-scope for self-closing elements.

fixes: #18335

Signed-off-by: Rémi Denis-Courmont <remi@remlab.net>