demux/playlist: xspf: only use text-elements inside current tag
These changes adds logic to not use an out-of-text text-element when
parsing an upcoming tag, as the text-element certainly is not part of
a tag that has not yet been parsed.

More accurately, a text-node that exists prior to an element that
has no text-body would previously inaccurately be treated as if it was
inside said element.

fixes: #18336

Signed-off-by: Rémi Denis-Courmont <remi@remlab.net>