demux/playlist: xspf: remove redundant comments
These comments does nothing to improve the comprehension of the
implementation as it is rather obvious what the lines that follow does
given the context in which they are in.

If someone has trouble understanding the semantics related to
XML_READER-enumerators, that is to be documented by the xml-reader
(not something that uses them).

These changes also removes misleading legacy comments that are no
longer matching the implementation itself (such as stating that we
only care about TrackList elements when parsing a playlist node (which
is not accurate).

Signed-off-by: Rémi Denis-Courmont <remi@remlab.net>