vlc_fingerprinter: use size_t for index argument in callback
It does not make sense for the passed index to be negative, nor is the
current diagnostic related to comparision of different signedness very
nice to look at.

This fixes both issues by changing the argument to size_t, making sure
that the values passed are suitable for interaction with
vlc_array-related functions.

Signed-off-by: Rémi Denis-Courmont <remi@remlab.net>