gui/qt: menus: always enable bookmarks menu
Due to the changes introduced by 36e12797b4a, the bookmark menu would
not be enabled if there were no choices currently associated with the
input. As a user should be able to manage the bookmarks, even if no
such is present at the current time, we forcefully add
VLC_VAR_HASCHOICE to the i_type.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>