input/control: UpdateBookmarksOption: prevent exessive bookmark-options
If UpdateBookmarksOption is called more than once, where the generated
bookmark-variable string would be different, it will result in the
input_item_t having several entries for "bookmarks=".

This patch makes it so that we overwrite any existing entry if such is
available, instead of simply appending a new one.

fixes: #18338

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>