playlist: playlist_NodeDelete: privatize forceful delete
Only the core should have permission to do forceful deletion of
entities.

This patch removes a mostly static, from outside of the core, argument
to playlist_NodeDelete, while also making it safer to use as it was
previously possible for anything with access to the function to delete
read-only entities.

Signed-off-by: Rémi Denis-Courmont <remi@remlab.net>