playlist: conditionally stop playlist on deletion of node
This fixes a regression introduced by 1ab80a3, where playback would
not stop if a node was deleted through user-interaction.

As a solution, a new flag to playlist_NodeDeleteExplicit has been
introduced so that we can conditionally stop playback of a deleted
item if desired (there are cases where we do not want such mechanisms,
such as when dealing with deletions of empty nodes).

fixes: #18350

Signed-off-by: Rémi Denis-Courmont <remi@remlab.net>