gui/qt: AddonsListModel::flags: fix disabling of addon
In the affected path we would like to toggle so that the addon is in a
state where it is not enabled by removing that flag (if present), we
would however remove all flags of the addon as the right-hand side
expression will always evaluate to 0.

This fixes the issue by proper use of bitwise negate.

Signed-off-by: Rémi Denis-Courmont <remi@remlab.net>