input: InputGetExtraFilesPattern: include scheme in concat-list entries
This allows for easier implementation for modules using the
concat-list, as it can assume that all entries can be passed directly
to functions such as vlc_stream_NewURL (as they are now valid URIs).

These changes also makes it so that the contents of concat-list
matches that of the documentation, as it uses the notion "input URLs".

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>