macOS: Unify VLCHUDToggleButton/VLCHUDImageButton to VLCImageButton
The implementation of these buttons is nearly the same, so it makes no
sense to have separate classes for them.

Only reason to do so, was the inability to get the button type, which
for now is worked-around by using an IBInspectable BOOL that has to be
set, if the button is a toggle button.

Not the nicest solution, but there is no better way, as far as I know.