macOS: Clarify useless-looking code in VLCDocumentController
Clarify why the VLCDocumentController is allocated and initialized in
the load method, so other developers are not confused by or accidentally
delete it, as it looks useless.
Additionally do not assign to an unused variable but just cast the
result to void, to make clear it is unused on purpose.