misc: ensure vlc_object_t is correctly aligned
To hide private data before (and/or after) public data, we normally
embed the public data within the private data structure. This ensures
proper alignment.

For VLC objects, that was unfortunately not an option, since the type of
the public data varies with the VLC object sub-classes. Thus the size,
and potentially the alignment, also vary.

This change ensures that the private data have the maximum possible
normal alignment. That will be adequate regardless of the public data
type.