macosx: Store subtitle settings from main menu in playlist scope
With main menu, the user shall only be able to modify settings
for the currently running VLC instance. All changes shall be
applied equally for all videos, therefore the variables are
created on playlist scope.

This avoid saving those changed to config, but still have them
valid for the overall VLC lifetime.

refs #16185