dbus: start interface before hand-off (fixes #3369)
This ensures that one instance gets the bus name before any instance
checks if the bus name is held. This fixes a race condition going all
the way back to the original VLC D-Bus interface implementation:
if multiple files are opened at the same time while no instance exists,
the new instances check the bus name for an owner before any of them
has had time to actually request and obtain the bus name. Then each
instance assumed it was the single one.

Note that there is still a race if the user closes the main instance
while trying to open files. Thits is an instrinsic user interface race,
and is unfixable.