macOS: Add VLCTintedImageButtonCell
This is a class for macOS versions lower than 10.10, where some image
buttons on dark backgrounds (or other colored backgrounds) would draw in
a dark color (usually black, for template images).
With this class, a tint color can be set for these buttons, it defaults
to white. This class is only intended to be used with template images.