marq: remove wrong x offset
This value is used as an unsigned in the vout and therefore causes an overflow
when the alignement contains SUBPICTURE_ALIGN_LEFT, which as the result of
having a text displayed at 426xxx in x. This value is supposed to be the
relative position from the alignement, and should therefore be 0 (as it always
is when opening the module and retrieving the alignement).

Signed-off-by: Thomas Guillem <thomas@gllm.fr>