misc/filter: remove assert from filter_AddProxyCallbacks
When a filter fails to load we will try to prepand a converter and then try to
load it again (using video_chroma/chain.c). The filter object will be reused
for the chain, therefore if any variable is created in the pf_activate of the
filter, it will exist in both the chroma chain object and the filter object
(because we do not require a module to destroy its variables). This assert will
then fail.

Signed-off-by: Thomas Guillem <thomas@gllm.fr>