hw: vaapi: remove vaapi global instance
We still need a *not global* vaapi instance: struct vlc_vaapi_instance. This
struct holds a VADisplay and a refcount. Every pictures of the vout pool hold
a reference to this struct in order to avoid that the VADisplay is terminated
while being used by pictures.

Filters now get the vaapi instance via filter_NewPicture(). Static filters like
the deinterlace one may not have access to a picture allocated by the vout.
This is fixed by next patches.

Refs #18445