macOS build.sh: blacklisting for compat with older macOS versions
VLC should be compiled with the latest SDK version to benefit from
new features. The currently latest SDK (10.12) already marks
availability for lots of libc functions. Unfortunately, as of now
the autotools configure checks do not support these annotations.
Therefore, lets disable known symbols which are not present on our
minimal macOS version, which is 10.7, currently.

refs #18476