Makefile: remove --disable-httpd
Unlike, say stream output, the HTTP server has no hooks anywhere within
the rest of the core: it is used exclusively through calls from modules
to the LibVLC run-time.

Thus, if no modules use it, the code will be eliminated by linker (in
static builds) or not paged in (in dynamic builds). There is not much
point in adding a build-time option to turn it off. It only creates
problems with keeping missing.c in sync.