macOS build.sh: Enable -Werror=partial-availability for contribs
Usually, contribs do not support runtime handling of partial
availability well. Turning this warning into an error has several
advantages:
- It fixes feature detection for custom build scripts: If a custom
build script includes the correct header, the check will now fail
to build and replacements will be enabled. This is the case for
ffmpeg, for instance.
- Even if configure check might not fail, the build will fail
for correctly annotated functions, once they are used in the code.
This makes it very visible if new partial symbols are added, so
that they can be fixed or disabled subsequently.

refs #18476