splitter: wall: fix incorrect sizes
The computations must not be done on the buffer size but on the video one.
Otherwise, we might end up with the sum of all output sizes bigger than the
video itself, which results in a green strip at the bottom of the lowest
outputs, and potentially incorrect display for all of them (with bad width).

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>