vout: rename and invert has_hide_mouse
The name was meant to indicate that the plugin hid the pointer
autonomously. But it was ostensibly confused as implying that the
plugin had a mouse hiding control (basically the exact opposite).
In fact, it really meant that the mouse hiding control would NOT be
used.

This tries to provide a slightly less confusing name, and accordingly
inverts the meaning. This is also realigns with the convention that
false should be the default setting (which was broken in the grand
grand parent of this changeset).