display: always compute mouse hide timeout
Cursor management should eventually be moved from display to the window.
(The video output core still needs to take care of the window to video
translation, and the mouse event filtering, as it already does.)
And cursor hiding should be left entirely within the window plugins.

But in the mean time, compute the correct timestamps regardless of the
cursor hiding support in display plugin, so that it also works if the
cursor is hidden by the window plugin.