display: add more documentation for display_info.handle_mouse_hide
In all cases this flag means that the vout has a custom Window and handles the
mouse event itself, including showing/hiding the mouse.

For the auto-hiding to work such vout need to call either
vout_display_SendEventMouseMoved or vout_display_SendEventMouseState to notify
the core of mouse events. And they all do that.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>