avcodec: favor D3D11VA over DXVA2
The order in which the vout are tried depends on which is the best for the
first video acceleration. Since we want D3D11 when possible we need to have
the matching VA by default too. On systems where D3D11VA doesn't work it will
fallback normally to DXVA2 and will use D3D9 with it.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>