display: ignore sample aspect ratio change failure
If display size change can require a resize without failure, then
changing the aspect ratio also ought to work.

In any case, the core never really did anything on failure.
In particular, it could not prevent the value change of the video
output variable, and did not perform any scaling.