Revert "input: De-duplicate titles"
This reverts commit 66401825ca4ba6f0d3c7d45d141bc15c4455506d.

This triggers a race between the interface and the input thread.
The playlist signals that a new input_thread_t is now the current item,
while the input thread tries to create the input_source_t containing the
titles.
If the interface reacts to the new item and tries to refresh the titles
before the input_source_t has been assigned to
input_thread_private_t::master, a crash will occur.
IMHO, having the titles duplicated is best for now, considering that
there will be massive changes to the playlist in 4.0