macosx: Fix filtering of log messages
Changes through the array controller are for the arrangedObjects, which
is the result of filtering and sorting.

Therefore we need to change the backing array. It is actually fine to do
that (contrary to my previous statement), as long as we behave correctly
and immediately send the appropriate KVO messages. Thats is why the
previous approach with just calling rearrangeObjects at some later point
after modifying the backing array would not work.

A big thanks to the AppKit Abusers Slack group, they provided a lot of
help to solve this.