macosx: audio effects: Convert constraints with firstBaseline
FirstBaseline was only introduces in 10.11 and crashes when running
on 10.7. Instead, use baseline constraints, which are fine for
single line controls.

BE CAREFUL: Xcode and ibtool compiler do NOT warn when you use
firstBaseline. Do not use the feature to auto add (missing)
constraints. This will add such firstBaseline contraints.