macosx: Use specific class for log messages
Instead of a NSDictionary, use a dedicated class for log messages.
Not only allows this for better abstraction and reusability, it
additionally solves a problem where log messages with the same content
would be considered equal, because all keys in the dictionary were
equal, ending up in wrong selections in the table.