macosx: Fix ignoring arguments in application:openFiles:
In the past, it was relied on the playlist item count to ignore items
that were already parsed by libvlc, but this is not enough as with some
command line flag combinations, like
  -v --video-splitter clone --clone-count 5
libvlc would not add any files, but Cocoa would try to open "clone" as
a file.

Therefore we need to ignore the intersection of the launch arguments and
the files to open.