libvlc: add an option to disable lua
Playlist lua modules are probed for each opened input. This can take a lot of
time, it can double the time of a local parsing where no lua module are needed.
This time is negligible when playing a file.

Playlist lua probe can take 5ms on desktop (in comparison of 5ms for parsing),
or 50ms on some android devices.

This new option can be used by VLC ports to disable lua when using VLC for
parsing.