httpd: simplify/fix connection closing
Close the connection if request is HTTP/1.0 or if the connection header
line in the response exists and is "close". This works because close is
the only that ever gets set as response so far; it wouldn't work if any
other token were emitted by the server.

Note: the RTSP server and the Lua art callback do not handle connection
close correctly still.