configure: add --with-libfuzzer
DIR is the dir path of the libFuzzer.a static library.

When used, it forces a static build without vlc binary and enable sanitize
coverage.

How to fetch and build libFuzzer.a from llvm.org:
$ git clone https://git.llvm.org/git/compiler-rt
$ cd compiler-rt/lib/fuzzer

for gcc:
$ CXX=g++ ./build.sh

for clang:
$ ./build.sh

This a partial merge of the work done by Shaleen during GSOC 2017. See
https://code.videolan.org/GSoC2017/shalzz/vlc.git