direct3d11: adjust the max luminance when tone mapping to SDR
The content never reaches 10000 nits also the max luminance in files is often
bogus, HDR TVs don't actually use them but use a hardcoded value instead.

When the display is in HDR mode, let it deal with the luminance. Otherwise
use a value of 5000 that gives a good result with most HDR10 content.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>