picture: remove dysfunctional picture_IsReferenced()
Nowadays, multiple pictures might share the same buffers. Consequently
the reference count of a single picture may be one even though there
are other references to the same buffers through a different picture.

Luckily picture_IsReferenced() is only really used in one place in
stream output, and then it was only an optimization.