chroma: copy: add tests
Test most common YUV420 conversions with SSE and without SSE. These conversions
are used via VAAPI, direct3d* and via videotoolbox.

There is no warranty that the source should be aligned when using copy.h, so
all tests are performed with a non aligned picture source. The destination is
always a picture allocated by VLC (so aligned).

The test verify that there are no memory corruption (via asan or just crashing)
and verify that colors are preserved. It will now assert if someone introduce a
green line.