codec: videotoolbox: pace output frames allocation
VT is not pacing frame allocation. If we are not fast enough to render
(release) the output pictures, the VT session can end up allocating way too
many frames. This can be problematic for 4K 10bits. To fix this issue, we
ensure that we don't have too many output frames allocated by waiting for the
vout to release them.

A proper way to fix this issue is to allow decoder modules to specify the dpb
and having the vout re-allocating output frames when this number changes.