demux: adaptive: fix misordered output
For packets like this:
A1, A2, B1, B2, A3...
which B1 has invalid timestamp and B2 has timestamp < barrier, B2 will
output but B1 is not, so misordered. A packet with invalid timestamp
should not output only if a packet with the same ES id has showed before
and timestamp larget than the barrier.

Signed-off-by: Francois Cartegnie <fcvlcdev@free.fr>