contribs: ffmpeg: work around the ffmpeg configure check for libmp3lame.
On Linux, libmp3lame.a requires -lm to be linked as well.  Now this
library does not have a pkg-config file, so there are not many options
to fix it.  One option is to pass extra-libs in our contrib system, and
another one is fixing libav/ffmpeg configure to try and link libmath
during the check.  Both are not good.

(cherry picked from commit d58a64505dd5488b328201ac6ea11919055279e9)
Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>