macosx: audio effects: Explain reasoning for UI manipulation in init
In init, the window is not loaded yet, so changing UI elements
does not work yet (pointers are null). It does not harm either, due to
objc behaviour on null object. But it also initializes the needed playlist
variables.
In the longer term, this should probably be splitted, as touching
lots of non-existing UI elements in init is not ideal.