qt(wayland): fix video widget failing to display video.
Setting Qt:WA_DontCreateNativeAncestors in addition to Qt::WA_NativeWindow
tells qt to create a native window for the widget and only for the widget.

For Wayland, this resulted in one single wayland subsurface instead of a
tree of nested subsurfaces. Wayland is a bit convoluted when it comes to
mapping/unmapping nested subsurfaces, which accounted for the video failing
to be displayed.

For X11, on the contrary, setting Qt:WA_DontCreateNativeAncestors proved
wrong with some misplacement in the video. So, this parameter is set
*** ONLY *** in a Wayland context.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>
(cherry picked from commit eb375bf9583eed0a1a588ee9d436a2219e2809f8)
Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>