video_output: reinit the vout if the crop changes
If you play a file that has the same decoder size but different display size
consecutively, the decoder picture pool needs to be updated so that the
i_visible_lines of each picture plane matches the new picture format.

The easiest way is to reinit the vout. Another way would be to update the
pictures in each picture pool dynamically. It cannot just be handled on the
vout side as some decoders and filters rely on i_visible/offset to copy
only the required data (for example lavc_CopyPicture).

Fixes #19268

Signed-off-by: Thomas Guillem <thomas@gllm.fr>